+++ bot.py
@@ -7,7 +7,7 @@ config = configparser.ConfigParser()
 config.read("config.ini")
 
 # Set up the bot.
-bot_description = """ Migaga (Version 4.0.1) """
+bot_description = """ Migaga (Version 4.1.0) """
 prefix = "!"
 client = commands.Bot(command_prefix=prefix, description=bot_description, pm_help=None)
 
@@ -30,6 +30,8 @@ extensions = [
 async def on_ready():
     print('Logged in as: ' + client.user.name)
     print('------')
+
+    setattr(client, "client_id", config.get("Env", "ClientId"))
 
     await client.change_presence(
         status=discord.Status.online, activity=discord.Game(